backend/requirements.txt
# Django
Django>=5.0.0,<6.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-environ>=0.11.0
django-filter>=24.0.0

# Database
psycopg2-binary>=2.9.9

# Authentication
djangorestframework-simplejwt>=5.3.0
djangorestframework-simplejwt[token_blacklist]>=5.3.0
django-allauth>=0.57.0

# Multi-tenant
django-tenants>=3.5.0

# File handling
Pillow>=10.2.0
django-storages>=1.14.2

# Internationalization
django-modeltranslation>=0.18.12

# Utilities
python-decouple>=3.8
celery>=5.3.4
redis>=5.0.1

# Development
black>=24.1.1
flake8>=7.0.0
pytest>=7.4.4
pytest-django>=4.8.0
pytest-cov>=4.1.0

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0

# Payment Gateways
stripe>=7.0.0
requests>=2.31.0  # Para Asaas (não tem SDK oficial)

